moved the set_device to the utils.py function and deleted device_util.py
