+++ avae/train.py
@@ -9,9 +9,9 @@ from sklearn.neighbors import KNeighborsClassifier
 
 from . import config, vis
 from .data import load_data
-from .device_util import set_device
 from .loss import AVAELoss
 from .model_b import AffinityVAE
+from .utils import set_device
 
 
 def train(
